rename module block_tensor -> block_sparse
